* Removed the LuaCob script, and redirected Cob2Lua calls to LuaRules * Removed the unused CommandDesc 'hotkey' member * Added the Spring.GetUnitIsDead() call-out * Made Spring.GetPositionLosState() backwards compatible * Added 'showPanelLabel = false' to layout.lua * Removed some 0.75b2 lua compatibility code
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@5333 37977431-3df6-0310-b722-df95706aa16b

+++ rts/System/Net/ProtocolDef.cpp
@@ -63,7 +63,7 @@ unsigned ProtocolDef::IsComplete(const unsigned char* const buf, const unsigned 
 	{
 		if (HasFixedLength(buf[0]))
 		{
-			if (bufLength >= (unsigned)GetLength(buf[0]))
+			if (bufLength >= (unsigned int)GetLength(buf[0]))
 				return GetLength(buf[0]);
 			else
 				return 0;
@@ -83,10 +83,10 @@ unsigned ProtocolDef::IsComplete(const unsigned char* const buf, const unsigned 
 				if (bufLength <= 2)
 					return 0;
 				
-				var = *(unsigned int*)(buf+1);
+				var = *((unsigned short*)(buf + 1));
 			}
 			
-			if (bufLength >= (unsigned)var)
+			if (bufLength >= (unsigned int)var)
 				return var;
 			else
 				return 0;
@@ -94,4 +94,4 @@ unsigned ProtocolDef::IsComplete(const unsigned char* const buf, const unsigned 
 	}
 }
 
-}
+} // namespace netcode
